requirements.txt
python-docx
PyMuPDF
python-pptx
sentence-transformers
chromadb
llama-cpp-python